workcortex-gmail-intelligence/requirements.txt
streamlit>=1.25.0
pandas>=1.3.0
openpyxl>=3.1.0
scikit-learn>=0.24.0
google-auth>=2.0.0
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
google-api-python-client>=2.50.0
numpy>=1.21.0
dnspython>=2.0.0
